mcux: Update to MCUXpresso SDK 2.7.0
Updates the MCUXpresso SDK to version 2.7.0 (2019-12-19) for all
available SoCs. The KL25Z and KW SoCs are not updated here because they
were not included in the MCUXpresso SDK 2.7.0 release.

Origin: MCUXpresso SDK
License: BSD 3-Clause
URL: mcux.nxp.com
Maintained-by: External

+++ mcux/drivers/imx/fsl_ewm.h
@@ -160,7 +160,7 @@ void EWM_GetDefaultConfig(ewm_config_t *config);
  */
 static inline void EWM_EnableInterrupts(EWM_Type *base, uint32_t mask)
 {
-    base->CTRL |= mask;
+    base->CTRL |= (uint8_t)mask;
 }
 
 /*!
@@ -175,7 +175,7 @@ static inline void EWM_EnableInterrupts(EWM_Type *base, uint32_t mask)
  */
 static inline void EWM_DisableInterrupts(EWM_Type *base, uint32_t mask)
 {
-    base->CTRL &= ~mask;
+    base->CTRL &= (uint8_t)(~mask);
 }
 
 /*!
@@ -195,7 +195,7 @@ static inline void EWM_DisableInterrupts(EWM_Type *base, uint32_t mask)
  */
 static inline uint32_t EWM_GetStatusFlags(EWM_Type *base)
 {
-    return (base->CTRL & EWM_CTRL_EWMEN_MASK);
+    return ((uint32_t)base->CTRL & EWM_CTRL_EWMEN_MASK);
 }
 
 /*!
